[#15] Changed the connector code to correctly use stoll() instead of stoi() when populating resource id.  This is evident in cockroachdb because the identifiers are 64 bit.

+++ lustre_irods_connector/src/irods_ops.cpp
@@ -120,7 +120,7 @@ int lustre_irods_connection::populate_irods_resc_id(lustre_irods_connector_cfg_t
     }
 
     try {
-        config_struct_ptr->irods_resource_id = std::stoi(&(resource_ids->value[0]));
+        config_struct_ptr->irods_resource_id = std::stoll(&(resource_ids->value[0]));
     } catch (std::invalid_argument& e) {
         clearGenQueryInp(&gen_inp);
         freeGenQueryOut(&gen_out);
